requirements.txt
certifi==2020.6.20
click==7.1.2
dnspython==1.16.0
dominate==2.4.0
ecdsa==0.13
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
Jinja2==2.11.2
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.18.2
numpydoc==0.9.1
psycopg2-binary==2.8.6
pytz==2020.1
PyYAML==5.3.1
sklearn==0.0
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
